Removed old ecal mapping and replaced it with a new functional version. Also fixed a bug in the analysis code when running with external timestamps.
